Implement API for cancelling CEF queries in JS
The query() function is replaced with a Query class, which allows the
request ID of a query to be stored and used to cancel the query. It's
slightly awkward because the ID can only be set when the query is sent.

A static send() method is used for queries that do not need to be
cancelled. For cancelable queries, the idea is that their object will be
stored in a container and their ID given to a UI cancel button, so that
clicking the button can lookup the object and send the cancel request.